requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
feedparser==6.0.11
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
PyYAML==6.0.2
sgmllib3k==1.0.0
yarl==1.18.3
